+++ src/constants.js
@@ -1,3 +1,4 @@
 export const FIRST_LEVEL_NAVIGATION_ID = 'ALL';
 export const MODULE_ID_ITEM_OVERVIEW = 'item-overview';
+export const MODULE_ID_SETTINGS = 'settings';
 export const MODULE_ID_WISHLIST = 'wishlist';