+++ src/containers/stages/header/SettingsIconContainer.js
@@ -1,0 +1,10 @@
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import setActiveModule from '../../../actions/navigation/setActiveModule';
+import SettingsIcon from '../../../components/stages/header/SettingsIcon';
+
+const mapDispatchToProps = (dispatch) => {
+    return bindActionCreators({ setActiveModule }, dispatch);
+};
+
+export default connect(undefined, mapDispatchToProps)(SettingsIcon);